Finished integration of SystemView. Required modifying the signature of _write_r in a SEGGER file, but that's okay. Tested functional

+++ segger_system_view/SEGGER/SEGGER_RTT_Syscalls_GCC.c
@@ -58,7 +58,7 @@
 ---------------------------END-OF-HEADER------------------------------
 File    : SEGGER_RTT_Syscalls_GCC.c
 Purpose : Low-level functions for using printf() via RTT in GCC.
-          To use RTT for printf output, include this file in your 
+          To use RTT for printf output, include this file in your
           application.
 Revision: $Rev: 9599 $
 ----------------------------------------------------------------------
@@ -76,7 +76,7 @@ Revision: $Rev: 9599 $
 **********************************************************************
 */
 //
-// If necessary define the _reent struct 
+// If necessary define the _reent struct
 // to match the one passed by the used standard library.
 //
 struct _reent;
@@ -88,7 +88,7 @@ struct _reent;
 **********************************************************************
 */
 int _write(int file, char *ptr, int len);
-int _write_r(struct _reent *r, int file, const void *ptr, int len);
+int _write_r(struct _reent *r, int file, const void *ptr, size_t len);
 
 /*********************************************************************
 *
@@ -123,7 +123,7 @@ int _write(int file, char *ptr, int len) {
 *   including stdout.
 *   Write data via RTT.
 */
-int _write_r(struct _reent *r, int file, const void *ptr, int len) {
+int _write_r(struct _reent *r, int file, const void *ptr, size_t len) {
   (void) file;  /* Not used, avoid warning */
   (void) r;     /* Not used, avoid warning */
   SEGGER_RTT_Write(0, ptr, len);
